Spring cleaning (#206)
Cleans up various changes that have been either been merged to vanilla or made redundant
Adds tooling to detect such issues in the future
Revamped changelogs

+++ createBuild.sh
@@ -7,6 +7,10 @@ fileHooksPath="src/lua/CompMod_FileHooks.lua"
 licenseFile="LICENSE"
 readMeFile="README.md"
 
+test $1 || { echo "Usage: $0 ns2_install_path"; exit 1; }
+install_path="$1"
+shift
+
 # Attempt to extract revision numbers from Filehooks file
 current_revision="$(cat $fileHooksPath | grep -oP 'g_compModRevision = \K[0-9]+')"
 current_beta_revision="$(cat $fileHooksPath | grep -oP 'g_compModBeta = \K[0-9]+')"
@@ -17,12 +21,16 @@ test -z "$current_beta_revision" && { echo "Failed to lookup current beta revisi
 revision_string="$current_revision"
 test "$current_beta_revision" -eq 0 || revision_string="$revision_string beta $current_beta_revision"
 
-# Check for outstanding commits
-test -n "$(git status --porcelain)" && { echo "ERROR: You have outstanding commits, please commit before creating a build"; exit 1; }
-
 echo "Creating build..."
 echo
 echo "CompMod Revision: $revision_string"
+
+# Check for outstanding commits
+test -n "$(git status --porcelain)" && { echo "ERROR: You have outstanding commits, please commit before creating a build"; exit 1; }
+
+# Build checks
+/bin/bash check_for_outdated_vars.sh "$install_path"
+test $? || { echo "ERROR: Build checks failed"; exit 1; }
 
 # Re-create the build dir
 test -d "$buildDir" && rm -rf "$buildDir"
